Move dispatch out to its own file
